+++ src/api/projects/labels.rs
@@ -9,8 +9,10 @@
 //! These endpoints are used for querying project labels.
 
 mod create;
+mod delete;
 mod label;
 mod labels;
+mod promote;
 
 pub use self::create::CreateLabel;
 pub use self::create::CreateLabelBuilder;
@@ -20,3 +22,9 @@ pub use self::label::LabelBuilder;
 
 pub use self::labels::Labels;
 pub use self::labels::LabelsBuilder;
+
+pub use self::delete::DeleteLabel;
+pub use self::delete::DeleteLabelBuilder;
+
+pub use self::promote::PromoteLabel;
+pub use self::promote::PromoteLabelBuilder;