+++ CHANGELOG.md
@@ -7,6 +7,10 @@
     `api::projects::protected_tags::ProtectedTag`
     `api::projects::protected_tags::ProtectedTags` endpoint to query, protect
     and unprotect a projects tags.
+  * Added the `api::projects::labels::DeleteLabel` endpoint to delete existing
+    labels from a project.
+  * Added the `api::projects::labels::PromoteLabel` endpoint to promote a project
+    label to a group label.
 
 # v0.1301.1